requirements.txt
crewai>=0.58.0
crewai-tools>=0.13.0
langchain>=0.2.14
langchain-openai>=0.1.20
pydantic>=2.7.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0.30
pandas>=2.2.2
matplotlib>=3.9.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
sqlglot>=25.14.0 # opcional, validación SQL